+++ packages/fiber/tests/web/web.core.test.tsx
@@ -93,6 +93,11 @@ describe('web core', () => {
     })
 
     expect(scene.children[0].type).toEqual('Mesh')
+    expect((scene.children[0] as ComponentMesh).geometry.type).toEqual('BoxGeometry')
+    expect((scene.children[0] as ComponentMesh).material.type).toEqual('MeshBasicMaterial')
+    expect((scene.children[0] as THREE.Mesh<THREE.BoxGeometry, MeshStandardMaterial>).material.type).toEqual(
+      'MeshBasicMaterial',
+    )
   })
 
   it('renders an empty scene', async () => {
@@ -138,6 +143,9 @@ describe('web core', () => {
     expect(scene.children[0].children[0].type).toEqual('Mesh')
     expect((scene.children[0].children[0] as ComponentMesh).geometry.type).toEqual('BoxGeometry')
     expect((scene.children[0].children[0] as ComponentMesh).material.type).toEqual('MeshBasicMaterial')
+    expect(
+      (scene.children[0].children[0] as THREE.Mesh<THREE.BoxGeometry, MeshStandardMaterial>).material.type,
+    ).toEqual('MeshBasicMaterial')
   })
 
   it('renders some basics with an update', async () => {
@@ -235,6 +243,24 @@ describe('web core', () => {
     expect(scene.children[0].children.length).toBe(0)
   })
 
+  it('can attach a Scene', async () => {
+    let scene: Scene = null!
+    await act(async () => {
+      scene = render(
+        <hasObject3dMember>
+          <scene attach="attachment" />
+        </hasObject3dMember>,
+        canvas,
+      ).getState().scene
+    })
+
+    const attachedScene = (scene.children[0] as HasObject3dMember).attachment
+    expect(attachedScene).toBeDefined()
+    expect(attachedScene?.type).toBe('Scene')
+    // attaching is *instead of* being a regular child
+    expect(scene.children[0].children.length).toBe(0)
+  })
+
   describe('attaches Object3D children that use attachFns', () => {
     it('attachFns as strings', async () => {
       let scene: Scene = null!
@@ -274,10 +300,10 @@ describe('web core', () => {
           <hasObject3dMethods>
             <mesh
               attachFns={[
-                (mesh: Mesh, parentInstance: HasObject3dMethods) => {
+                (mesh: Mesh) => {
                   attachedMesh = mesh
                 },
-                (mesh: Mesh, parentInstance: HasObject3dMethods) => {
+                (mesh: Mesh) => {
                   detachedMesh = mesh
                 },
               ]}
@@ -404,7 +430,7 @@ describe('web core', () => {
       }
     }
 
-    await expect(async () => {
+    expect(async () => {
       await act(async () => {
         extend({ MyColor })