+++ packages/fiber/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @react-three/fiber
 
+## 7.0.10
+
+### Patch Changes
+
+- e019dd4: fixes
+
+## 7.0.9
+
+### Patch Changes
+
+- cd266e4: Fix diffProps dashed keys
+
+## 7.0.8
+
+### Patch Changes
+
+- 6f68406: Allow getCurrentViewport to receive an array
+
 ## 7.0.7
 
 ### Patch Changes